Add files via upload
This is the file with the code to create the form fields for the user to fill in and has a live preview. There is a loadshedding feature as an idea of which I thought of for learning purposes, meaning it has no logic yet but a button the user switches on to view the loadshedding status.